improved Readme.md, changed repository file structure
