+++ Makefile
@@ -1,4 +1,6 @@
+setup_requirements:
+	pip install -r requirements.txt
 set_institution_db:
-	python source/makePython/download_excel.py
-	python source/makePython/parse_excel.py
-	python source/Front-End/MainMenu/qtdropdown.py
+	python source/program/main_menu/helpers/download_excel.py
+	python source/program/main_menu/helpers/parse_excel.py
+	python source/program/main_menu/qtdropdown.py
