+++ run.bat
@@ -1,3 +1,3 @@
-python source/makePython/download_excel.py
-python source/makePython/parse_excel.py
-python source/Front-End/MainMenu/qtdropdown.py
+python source/program/main_menu/helpers/download_excel.py
+python source/program/main_menu/helpers/parse_excel.py
+python source/program/main_menu/qtdropdown.py
