+++ setup/setup.bat
@@ -1,0 +1,1 @@
+pip install -r requirements.txt
